pyproject.toml
[project]
name = "librepythonista-python-editor"
version = "0.1.6"
description = "A Python editor for LibreOffice that has the LibrePythonista extension installed."
readme = "README.md"
authors = [
    {name = ":Barry-Thomas-Paul: Moss", email = "4193389+Amourspirit@users.noreply.github.com"}
]
license = {file = "LICENSE"}
keywords = ["LibrePythonista", "LibreOffice", "ooodev", "extension"]
requires-python = ">=3.8"
# https://packaging.python.org/en/latest/specifications/dependency-specifiers/
dependencies = [
    "jedi>=0.19.0",
    "cefpython3>=66.1; sys_platform == 'win32'",
    "pyobjc>=10.3.1; platform_system == 'darwin'",
    "PyQt6>=6.7.0; platform_system == 'darwin'",
    "qtpy>=2.4.0; platform_system == 'linux'",
    "PyQt5>=5.15.4; platform_system == 'linux'",
    "PySide2>=5.15.0; platform_system == 'linux'",
    "pywebview>=5.3.2"    
]

classifiers = [

    # Indicate who your project is intended for
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",

    # Pick your license as you wish (see also "license" above)
    "License :: OSI Approved :: Apache Software License",

    # Specify the Python versions you support here.
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]


[project.urls]
Homepage = "https://github.com/Amourspirit/python-librepythonista-python-editor"
Repository = "https://github.com/Amourspirit/python-librepythonista-python-editor"
Issues = "https://github.com/Amourspirit/python-librepythonista-python-editor/issues"


[project.scripts]
librepythonista-cell-edit = "librepythonista_py_edit.cell_edit:main"

[dependency-groups]
dev = [
    "python-dotenv>=1.0.1",
]


# https://packaging.python.org/en/latest/guides/writing-pyproject-toml/

# [[tool.uv.index]]
# name = "pypi"
# url = "https://pypi.org/legacy/"
# default = true
# token = {env = "UV_TOKEN"}

[[tool.uv.index]]
name = "pypi-test"
url = "https://test.pypi.org/legacy/"
default = false
username= "__token__"
password = {env = "UV_TEST_TOKEN"}

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
    "librepythonista_py_edit/**/*",
]